+++ R/xml_write.R
@@ -12,7 +12,10 @@
 #'
 #' tmp <- tempfile(fileext = ".xml")
 #' write_xml(h, tmp)
-#' read_xml(tmp)
+#' readLines(tmp)
+#'
+#' write_html(h, tmp)
+#' readLines(tmp)
 write_xml <- function(x, file, ...) {
   UseMethod("write_xml")
 }
@@ -24,7 +27,7 @@ write_xml.xml_missing <- function(x, file, ...) {
 
 #' @export
 write_xml.xml_document <- function(x, file, format = TRUE, ...) {
-  doc_write(x$doc, file, format)
+  doc_write_xml(x$doc, file, format)
 }
 
 #' @export
@@ -33,10 +36,41 @@ write_xml.xml_nodeset <- function(x, file, ...) {
     stop("Can only save length 1 node sets", call. = FALSE)
   }
 
-  node_write(x[[1]]$node, x$doc, file, ...)
+  node_write_xml(x[[1]]$node, x$doc, file, ...)
 }
 
 #' @export
 write_xml.xml_node <- function(x, file, format = TRUE, ...) {
-  node_write(x$node, x$doc, file)
+  node_write_xml(x$node, x$doc, file)
+}
+
+
+#' @export
+#' @rdname write_xml
+write_html <- function(x, file, ...) {
+  UseMethod("write_html")
+}
+
+#' @export
+write_html.xml_missing <- function(x, file, ...) {
+  stop("Missing data cannot be written", call. = FALSE)
+}
+
+#' @export
+write_html.xml_document <- function(x, file, format = TRUE, ...) {
+  doc_write_html(x$doc, file, format)
+}
+
+#' @export
+write_html.xml_nodeset <- function(x, file, ...) {
+  if (length(x$nodeset) != 1) {
+    stop("Can only save length 1 node sets", call. = FALSE)
+  }
+
+  node_write_html(x[[1]]$node, x$doc, file, ...)
+}
+
+#' @export
+write_html.xml_node <- function(x, file, format = TRUE, ...) {
+  node_write_html(x$node, x$doc, file)
 }